Add files via upload
This project I worked on real-world dataset from telecom company that includes customer information such as demographics, usage patterns, and whether they churned or not. The goal is to prepare that data for analysis, explore key pattern and trends,  and create visualizations to gain insights into factor influencing customer churn.